Revert "Cleanups - remove/silence warnings"
This reverts commit cf04a2c0fe41b4c8428f0103a4f6e6b41270be1b.

+++ menu/drivers/ozone.c
@@ -8108,7 +8108,7 @@ static enum menu_action ozone_parse_menu_entry_action(
 
             if (     !menu_navigation_wraparound_enable
                   && selection == ozone->system_tab_end + horizontal_list_size)
-               new_selection   = (int)selection;
+               new_selection   = selection;
 
             if (new_selection != (int)selection)
             {
